requirements.txt
werobot==1.13.1
python-dotenv==0.21.0
requests==2.28.1
tencentcloud-sdk-python==3.0.772
stability-sdk==0.2.7
